Cv_project/requirements.txt
torch==2.0.1
diffusers==0.13.0
Pillow==9.3.0
scikit-learn==1.1.3
joblib==1.2.0
numpy==1.23.4
torchvision==0.14.0
